Firsttwoscreensconceptualize
So far I have added a scroll and table layout and have begun to
organize elements on the screen. Right now I am using an imageView to
display the image, however, I plan on changing it to an imageButton so
I can use it as a navigational element to go to the media gallery for
photos and videos.